move `Table` from structures to bricks
