+++ .changeset/spotty-ties-poke.md
@@ -1,0 +1,10 @@
+---
+"@stratakit/structures": patch
+---
+
+The `Table` component has been moved from `@stratakit/structures` to `@stratakit/bricks` and marked as deprecated. Please update your imports:
+
+```diff
+- import { Table } from "@stratakit/structures";
++ import { Table } from "@stratakit/bricks";
+```
