+++ apps/test-app/app/~meta.ts
@@ -20,6 +20,7 @@ export const components = {
 		"Skeleton",
 		"Spinner",
 		"Switch",
+		"Table",
 		"Text",
 		"TextBox",
 		"Tooltip",
@@ -34,7 +35,6 @@ export const components = {
 		"NavigationList",
 		"NavigationRail",
 		"Popover",
-		"Table",
 		"Tabs",
 		"Toolbar",
 		"Tree",
